Added useGameActions hook
